# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection",
        "title": "mindscape",
        "contact": {
            "name": "Adam Kalinowski",
            "url": "https://github.com/adamkali/mindscape"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/bookmarks": {
            "post": {
                "description": "Create a new Bookmark by Authorization Header",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Create a new Bookmark",
                "operationId": "CreateBookmark",
                "parameters": [
                    {
                        "description": "CreateBookmarkRequest",
                        "name": "CreateBookmarkRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repository.CreateBookmarkParams"
                        }
                    },
                    {
                        "type": "string",
                        "default": "\"Bearer token\"",
                        "description": "Authorization Header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BookmarkResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/BookmarkResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/BookmarkResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/BookmarkResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Move a Bookmark by Authorization Header, and my a\nParentFolderId [parent_id].",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Move a Bookmark",
                "operationId": "MoveBookmark",
                "parameters": [
                    {
                        "description": "Move Bookmark Request",
                        "name": "MoveBookmarkRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MoveBookmarkRequest"
                        }
                    },
                    {
                        "type": "string",
                        "default": "\"Bearer token\"",
                        "description": "Authorization Header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BookmarksResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/BookmarksResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/BookmarksResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/BookmarksResponse"
                        }
                    }
                }
            }
        },
        "/bookmarks/folder/{bookmark_id}": {
            "delete": {
                "description": "Delete a Bookmark by Authorization Header, and my a\nParentFolderId [parent_id].",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Delete a Bookmark",
                "operationId": "DeleteBookmark",
                "parameters": [
                    {
                        "type": "string",
                        "default": "\"Bearer token\"",
                        "description": "Authorization Header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"e38e78a4-2ca3-4c59-a3ea-a2019866e593\"",
                        "description": "Parent Folder ID",
                        "name": "bookmark_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BookmarksResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/BookmarksResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/BookmarksResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/BookmarksResponse"
                        }
                    }
                }
            }
        },
        "/bookmarks/folder/{parent_id}": {
            "get": {
                "description": "Get all Bookmarks by Authorization Header and by the\nParentFolderId [parent_id]",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Get Bookmarks By Folder ID",
                "operationId": "GetBookmarks",
                "parameters": [
                    {
                        "type": "string",
                        "default": "\"Bearer token\"",
                        "description": "Authorization Header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"e38e78a4-2ca3-4c59-a3ea-a2019866e593\"",
                        "description": "Parent Folder ID",
                        "name": "parent_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BookmarksResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/BookmarksResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/BookmarksResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/BookmarksResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "BookmarkResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/repository.Bookmark"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "BookmarksResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.Bookmark"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "MoveBookmarkRequest": {
            "type": "object",
            "properties": {
                "bookmarkId": {
                    "type": "string"
                },
                "newParentId": {
                    "description": "The parent folder to move the bookmark to cannot be null",
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "repository.Bookmark": {
            "type": "object",
            "properties": {
                "created_datetime": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "folder_id": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_datetime": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "repository.CreateBookmarkParams": {
            "type": "object",
            "properties": {
                "folder_id": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    }
}